// frontend/src/ts/types/font-awesome.d.ts
export type FaObject =
  | {
      variant: "solid";
      icon: FaSolidIcon;
    }
  | {
      variant: "regular";
      icon: FaRegularIcon;
    }
  | {
      variant: "brand";
      icon: FaBrandIcon;
    }
  | {
      variant?: undefined;
      icon: FaSolidIcon;
    };
// ...
export type FaRegularIcon =
  | "address-book"
  | "address-card"
  | "angry"
  | "arrow-alt-circle-down"
  | "arrow-alt-circle-left"
  | "arrow-alt-circle-right"
  | "arrow-alt-circle-up"
  | "bell"
  | "bell-slash"
  | "bookmark"
  | "building"
  | "calendar"
  | "calendar-alt"
  | "calendar-check"
  | "calendar-minus"
  | "calendar-plus"
  | "calendar-times"
  | "caret-square-down"
  | "caret-square-left"
  | "caret-square-right"
  | "caret-square-up"
  | "chart-bar"
  | "check-circle"
  | "check-square"
  | "circle"
  | "clipboard"
  | "clock"
  | "clone"
  | "closed-captioning"
  | "comment"
  | "comment-alt"
  | "comment-dots"
  | "comments"
  | "compass"
  | "copy"
  | "copyright"
  | "credit-card"
  | "dizzy"
  | "dot-circle"
  | "edit"
  | "envelope"
  | "envelope-open"
  | "eye"
  | "eye-slash"
  | "file"
  | "file-alt"
  | "file-archive"
  | "file-audio"
  | "file-code"
  | "file-excel"
  | "file-image"
  | "file-pdf"
  | "file-powerpoint"
  | "file-video"
  | "file-word"
  | "flag"
  | "flushed"
  | "folder"
  | "folder-open"
  | "font-awesome-logo-full"
  | "frown"
  | "frown-open"
  | "futbol"
  | "gem"
  | "grimace"
  | "grin"
  | "grin-alt"
  | "grin-beam"
  | "grin-beam-sweat"
  | "grin-hearts"
  | "grin-squint"
  | "grin-squint-tears"
  | "grin-stars"
  | "grin-tears"
  | "grin-tongue"
  | "grin-tongue-squint"
  | "grin-tongue-wink"
  | "grin-wink"
  | "hand-lizard"
  | "hand-paper"
  | "hand-peace"
  | "hand-point-down"
  | "hand-point-left"
  | "hand-point-right"
  | "hand-point-up"
  | "hand-pointer"
  | "hand-rock"
  | "hand-scissors"
  | "hand-spock"
  | "handshake"
  | "hdd"
  | "heart"
  | "hospital"
  | "hourglass"
  | "id-badge"
  | "id-card"
  | "image"
  | "images"
  | "keyboard"
  | "kiss"
  | "kiss-beam"
  | "kiss-wink-heart"
  | "laugh"
  | "laugh-beam"
  | "laugh-squint"
  | "laugh-wink"
  | "lemon"
  | "life-ring"
  | "lightbulb"
  | "list-alt"
  | "map"
  | "meh"
  | "meh-blank"
  | "meh-rolling-eyes"
  | "minus-square"
  | "money-bill-alt"
  | "moon"
  | "newspaper"
  | "object-group"
  | "object-ungroup"
  | "paper-plane"
  | "pause-circle"
  | "play-circle"
  | "plus-square"
  | "question-circle"
  | "registered"
  | "sad-cry"
  | "sad-tear"
  | "save"
  | "share-square"
  | "smile"
  | "smile-beam"
  | "smile-wink"
  | "snowflake"
  | "square"
  | "star"
  | "star-half"
  | "sticky-note"
  | "stop-circle"
  | "sun"
  | "surprise"
  | "thumbs-down"
  | "thumbs-up"
  | "times-circle"
  | "tired"
  | "trash-alt"
  | "user"
  | "user-circle"
  | "window-close"
  | "window-maximize"
  | "window-minimize"
  | "window-restore";